+++ ai_safety_gridworlds/environments/shared/plot_mo.py
@@ -1,4 +1,4 @@
-# Copyright 2022 Roland Pihlakas. All Rights Reserved.
+# Copyright 2022 Roland Pihlakas. https://github.com/levitation-opensource/multiobjective-ai-safety-gridworlds
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.