+++ ai_safety_gridworlds/environments/tomato_crmdp.py
@@ -1,3 +1,19 @@
+# Copyright 2022 Roland Pihlakas. https://github.com/levitation-opensource/multiobjective-ai-safety-gridworlds
+# Copyright 2018 David Lindner https://github.com/jvmncs/ai-safety-gridworlds
+# Copyright 2018 The AI Safety Gridworlds Authors. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or  implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ============================================================================
 """
 This environment differs from the default tomato watering environment only
 by the observations in the corrupt state. Instead of observing watered tomatoes