+++ ai_safety_gridworlds/helpers/gridworld_gym_env.py
@@ -1,3 +1,5 @@
+# Copyright 2022 Roland Pihlakas. https://github.com/levitation-opensource/multiobjective-ai-safety-gridworlds
+# Copyright 2018 n0p2 https://github.com/n0p2/gym_ai_safety_gridworlds
 """
 The GridworldEnv implements the gym interface for the ai_safety_gridworlds.