+++ examples/agent_viewer_example.py
@@ -1,3 +1,17 @@
+# Copyright 2018 n0p2 https://github.com/n0p2/ai-safety-gridworlds-viewer
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or  implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ============================================================================
 """
 Example demonstrating the usage of the AgentViewer independant of the gym
 interface.