+++ examples/gym_env_example.py
@@ -1,3 +1,4 @@
+# Copyright 2018 n0p2 https://github.com/n0p2/gym_ai_safety_gridworlds
 """
 Example demonstrating the usage of the Gym interface.